requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
attrs==25.4.0
blockbuster==1.5.26
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
cryptography==46.0.3
dataclasses-json==0.6.7
debugpy==1.8.19
decorator==5.2.1
executing==2.2.1
forbiddenfruit==0.1.4
frozenlist==1.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-tools==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
jupyter_client==8.7.0
jupyter_core==5.9.1
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.5
langchain-experimental==0.4.1
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-api==0.6.15
langgraph-checkpoint==3.0.1
langgraph-cli==0.4.11
langgraph-prebuilt==1.0.5
langgraph-runtime-inmem==0.20.1
langgraph-sdk==0.3.1
langsmith==0.5.1
marshmallow==3.26.2
matplotlib-inline==0.2.1
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.4.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
parso==0.8.5
platformdirs==4.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.45
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.50.0
structlog==25.5.0
tenacity==9.1.2
tornado==6.5.4
traitlets==5.14.3
truststore==0.10.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.40.0
watchfiles==1.1.1
wcwidth==0.2.14
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0


fastapi==0.115.5
uvicorn[standard]==0.34.0
pydantic==2.10.5
